requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiohttp_socks==0.9.1
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
frozenlist==1.5.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pycares==4.5.0
pycparser==2.22
Pygments==2.18.0
python-dotenv==1.0.1
pyTelegramBotAPI==4.25.0
python-socks==2.5.3
requests==2.32.3
rich==13.9.4
setuptools==75.6.0
tgtg==0.18.1
urllib3==2.2.3
wheel==0.45.1
yarl==1.18.3
